backend/requirements.txt
Flask==3.1.3
Flask-SocketIO==5.6.1
psycopg2-binary==2.9.11
bcrypt==4.1.3
python-dotenv==1.0.1
eventlet==0.36.1
Pillow==10.4.0
PyJWT